change_master: Show how to change master storage domain
Add example accepting data center and storage domain names and setting
the storage domain as the master storage domain.

The example show how to find the entities by name, and how to verify if
the storage domain is already the master storage domain.

Switching master storage domain is not possible currently when SPM tasks
are running. The example show how to detect the error and retry the
operation.

With this script and some manual work, I could reproduce the bug:
https://bugzilla.redhat.com/1913764#c35

See example run log here:
https://bugzilla.redhat.com/attachment.cgi?id=1822866

Related-To: https://bugzilla.redhat.com/1913764
Signed-off-by: Nir Soffer <nsoffer@redhat.com>