list_disk_snapshots.py: look for template snapshots in the chain
When a VM is created from template with Thin Storage Allocation,
the parent of one of the snapshots will be the template image, thus
not returned in the result set for the list disksnapshots request.
In this case, when requesting snapshots of a disk created this way -
this script returns an error.

This patch takes advantage of the disks/xxx/disksnapshots API with
'include_active' and 'include_template' flags to fetch all the
snapshots required for the chain reconstruction

Bug-Url: https://bugzilla.redhat.com/1900597