backup_vm: Pass the backup to download_backup()
download_backup() was checking verifying that the vm and the backup
exist, and verifying that the backup is ready. This should really be
done in cmd_download, since when calling download_backup() from
cmd_full() or cmd_incremental() we know the backup exists and ready.

Move the checks up to cmd_download(), and pass the backup object to
download_backup(). This will be useful later to get the checkpoint id
from the backup.