backup_vm: Include checkpoint id in backup filename
This will make it easier to locate the previous backup when creating an
incremental backup. To improve sorting, move the timestamp to the start
of the filename.

Here is a backup flow showing how the checkpoint is useful:

1. Create full backup

    $ ./backup_vm.py -c engine full 54fada0b-94c6-4374-8cf2-fe477c7fcd76 \
        --backup-dir /data/scratch/backups/
    ...

2. Find the last backup:

    $ ls /data/scratch/backups/
    20220303185549.542990c2-f2bd-4f8e-a5b8-c184c8aaabbb.85e58f90-745d-4a81-bc02-2e4550ca74d2.full.qcow2

3. Created incremental backup since this backup:

    $ ./backup_vm.py -c engine incremental 54fada0b-94c6-4374-8cf2-fe477c7fcd76 \
        --backup-dir /data/scratch/backups/ \
        --from-checkpoint-uuid 542990c2-f2bd-4f8e-a5b8-c184c8aaabbb
    ...

4. Find the last backup:

    $ ls -t /data/scratch/backups/
    20220303185913.a3414532-a2ed-4d62-9cdb-541671d01c2b.85e58f90-745d-4a81-bc02-2e4550ca74d2.incremental.qcow2
    20220303185549.542990c2-f2bd-4f8e-a5b8-c184c8aaabbb.85e58f90-745d-4a81-bc02-2e4550ca74d2.full.qcow2

With this change, we can easily find the checkpoint id for the next
incremental backup by sorting the backup files. This will be useful for
actually rebasing the incremental backup on top of the previews backup.