backup_vm: Pass the backup to stop_backup()
Verifying that vm exists and backup can be stopped should be done in
cmd_stop(), not in stop_backup(). When we call stop_backup() from
cmd_full() or cmd_incremental() we know that the vm exits and that the
backup is ready.

In a real application it would be good idea to skip stopping if the
backup has already succeeded, but in this script it is useful to send
stop request more than once for testing purposes.