Improve backup vm script (#60)
* backup_vm: Add seconds to backup timestamp

When testing backups, it is easy to create multiple backups in the same
minutes, overwriting older backup.

* backup_vm: Shorten clumsy progress message

Creating checkpoint xxx-yyy is good enough message. The previous text
about using the checkpoint in the next backup is useful for someone
using this script for the first time, but after that is just noise.

* backup_vm: Fix IndexError when backup fails

If backup fails without logging the expected audit log, backup_vm fails
with IndexError, hiding the original error. Fix by changing the helper
to return the last error instead of all errors, and return None if there
is no event.

* backup_vm: Fix logging on errors

If downloading a disk raised, we logged "Download completed
successfully" before raising the error, which is very confusing.
Now we log this message only when the download succeeded.

* backup_vm: Pass the backup to download_backup()

download_backup() was checking verifying that the vm and the backup
exist, and verifying that the backup is ready. This should really be
done in cmd_download, since when calling download_backup() from
cmd_full() or cmd_incremental() we know the backup exists and ready.

Move the checks up to cmd_download(), and pass the backup object to
download_backup(). This will be useful later to get the checkpoint id
from the backup.

* backup_vm: Pass the backup to stop_backup()

Verifying that vm exists and backup can be stopped should be done in
cmd_stop(), not in stop_backup(). When we call stop_backup() from
cmd_full() or cmd_incremental() we know that the vm exits and that the
backup is ready.

In a real application it would be good idea to skip stopping if the
backup has already succeeded, but in this script it is useful to send
stop request more than once for testing purposes.

* backup_vm: Include checkpoint id in backup filename

This will make it easier to locate the previous backup when creating an
incremental backup. To improve sorting, move the timestamp to the start
of the filename.

Here is a backup flow showing how the checkpoint is useful:

1. Create full backup

    $ ./backup_vm.py -c engine full 54fada0b-94c6-4374-8cf2-fe477c7fcd76 \
        --backup-dir /data/scratch/backups/
    ...

2. Find the last backup:

    $ ls /data/scratch/backups/
    20220303185549.542990c2-f2bd-4f8e-a5b8-c184c8aaabbb.85e58f90-745d-4a81-bc02-2e4550ca74d2.full.qcow2

3. Created incremental backup since this backup:

    $ ./backup_vm.py -c engine incremental 54fada0b-94c6-4374-8cf2-fe477c7fcd76 \
        --backup-dir /data/scratch/backups/ \
        --from-checkpoint-uuid 542990c2-f2bd-4f8e-a5b8-c184c8aaabbb
    ...

4. Find the last backup:

    $ ls -t /data/scratch/backups/
    20220303185913.a3414532-a2ed-4d62-9cdb-541671d01c2b.85e58f90-745d-4a81-bc02-2e4550ca74d2.incremental.qcow2
    20220303185549.542990c2-f2bd-4f8e-a5b8-c184c8aaabbb.85e58f90-745d-4a81-bc02-2e4550ca74d2.full.qcow2

With this change, we can easily find the checkpoint id for the next
incremental backup by sorting the backup files. This will be useful for
actually rebasing the incremental backup on top of the previews backup.

* backup_vm: Remove unneeded helper

To get backup disk backup_mode, we can access the backup disk
backup_mode attribute. There is no need to get the disk again from the
disks_service.

* backup_vm: Create qcow2 chain

When downloading an incremental backup, try to use the previous backup
file as a backing file. This creates a chain that can be used to
restore the disk later.

With this change, we don't need to use qemu-img rebase to rebase the
the incremental backup, which is hard to do correctly. This makes the
script more useful for testing or for creating simple backup solution.

Here is an example backup with this change:

1. Create a full backup

$ ./backup_vm.py -c engine full 54fada0b-94c6-4374-8cf2-fe477c7fcd76 --backup-dir backups
[   0.0 ] Starting full backup for VM '54fada0b-94c6-4374-8cf2-fe477c7fcd76'
[   0.7 ] Waiting until backup '2b80a286-eed3-4a39-91f4-21725050f104' is ready
[  12.8 ] Created checkpoint '35b51c4e-b8a4-4a2e-a412-9e443948c2b4'
[  12.8 ] Downloading full backup for disk '85e58f90-745d-4a81-bc02-2e4550ca74d2'
[  12.8 ] Creating backup file 'backups/20220303225543.35b51c4e-b8a4-4a2e-a412-9e443948c2b4...full.qcow2'
[  14.4 ] Image transfer 'e6ae86b2-6672-4fbe-834f-0dff58ce0f47' is ready
[ 100.00% ] 6.00 GiB, 2.82 seconds, 2.13 GiB/s
[  17.2 ] Finalizing image transfer
[  21.3 ] Download completed successfully
[  21.3 ] Finalizing backup
[  28.5 ] Full backup '2b80a286-eed3-4a39-91f4-21725050f104' completed successfully

2. Lookup the last backup

$ ls -t backups/
20220303225543.35b51c4e-b8a4-4a2e-a412-9e443948c2b4.85e58f90-745d-4a81-bc02-2e4550ca74d2.full.qcow2

3. Create incremental backup

$ ./backup_vm.py -c engine incremental 54fada0b-94c6-4374-8cf2-fe477c7fcd76 \
    --backup-dir backups \
    --from-checkpoint-uuid 35b51c4e-b8a4-4a2e-a412-9e443948c2b4
[   0.0 ] Starting incremental backup for VM '54fada0b-94c6-4374-8cf2-fe477c7fcd76'
[   0.5 ] Waiting until backup 'ee79ef1e-fc12-48d1-b899-8b8bec171b42' is ready
[  13.6 ] Created checkpoint '7eb81de9-fd0a-44fe-a997-42eca695a6a7'
[  13.7 ] Downloading incremental backup for disk '85e58f90-745d-4a81-bc02-2e4550ca74d2'
[  13.7 ] Creating backup file 'backups/20220303225724.7eb81de9-fd0a-44fe-a997-42eca695a6a7...incremental.qcow2'
[  13.7 ] Using backing file '20220303225543.35b51c4e-b8a4-4a2e-a412-9e443948c2b4...full.qcow2'
[  15.2 ] Image transfer 'f896d25e-e987-405f-9a4b-530bcb82c27d' is ready
[ 100.00% ] 6.00 GiB, 0.08 seconds, 75.91 GiB/s
[  15.3 ] Finalizing image transfer
[  17.3 ] Download completed successfully
[  17.3 ] Finalizing backup
[  20.5 ] Incremental backup 'ee79ef1e-fc12-48d1-b899-8b8bec171b42' completed successfully

4. The backup files

$ ls -t backups
20220303225724.7eb81de9-fd0a-44fe-a997-42eca695a6a7.85e58f90-745d-4a81-bc02-2e4550ca74d2.incremental.qcow2
20220303225543.35b51c4e-b8a4-4a2e-a412-9e443948c2b4.85e58f90-745d-4a81-bc02-2e4550ca74d2.full.qcow2

5. The backing chain

$ qemu-img info --backing-chain backups/20220303225724.7eb81de9-fd0a-44fe-a997-42eca695a6a7...incremental.qcow2
image: backups/20220303225724.7eb81de9-fd0a-44fe-a997-42eca695a6a7...incremental.qcow2
file format: qcow2
virtual size: 6 GiB (6442450944 bytes)
disk size: 196 KiB
cluster_size: 65536
backing file: 20220303225543.35b51c4e-b8a4-4a2e-a412-9e443948c2b4...full.qcow2
    (actual path: backups/20220303225543.35b51c4e-b8a4-4a2e-a412-9e443948c2b4...full.qcow2)
backing file format: qcow2
Format specific information:
    compat: 1.1
    compression type: zlib
    lazy refcounts: false
    refcount bits: 16
    corrupt: false
    extended l2: false

image: backups/20220303225543.35b51c4e-b8a4-4a2e-a412-9e443948c2b4...full.qcow2
file format: qcow2
virtual size: 6 GiB (6442450944 bytes)
disk size: 1.12 GiB
cluster_size: 65536
Format specific information:
    compat: 1.1
    compression type: zlib
    lazy refcounts: false
    refcount bits: 16
    corrupt: false
    extended l2: false